OpenAPI spec updated and used for encrypt status.
[ upstream commit 7cdadbcea9c9f48e717a670ab287b1933703705a ]

Global variable `countErrors` converted to the function local.

+++ cilium/cmd/encrypt_status_test.go
@@ -16,10 +16,10 @@ const procTestFixtures = "fixtures/proc"
 func (s *EncryptStatusSuite) TestGetXfrmStats(c *C) {
 	errCount, m, err := getXfrmStats(procTestFixtures)
 	c.Assert(err, Equals, nil)
-	currentCount := 0
+	currentCount := int64(0)
 	testCases := []struct {
 		name string
-		want int
+		want int64
 	}{
 		{name: "XfrmInError", want: 2},
 		{name: "XfrmInBufferError", want: 0},
